Provide partial Coverage #313
- add back Jacoco configuration
- now Sonar requires the xml report to be generated, it means that we
need to launch verify phase instead of test
- excluding 2 tests from coverage, they are playing and it is causing
Jacoco to not report any coverage 90% of the time. This look alike a bug
in Jacoco, will need to qualify and report it.

So this PR is adding back coverage of the 232 tests over the 238
existing. This is not yet perfect but better than 0.

Signed-off-by: Aurélien Pupier <apupier@redhat.com>