FUSETOOLS2-512 - first iteration to parse modeline
No impact on external behavior. It is the first piece to be able to
provide support for modeline. Now that there is a model for it we will
eb able to provide it. The model will surely evolve to handle completion
in Option Values but it can be done later to keep small chunks of code.

- parse into a model taking care of position with options
- handling modeline for all supported languages (yaml and xml are
special cases compared to others)
- keep same level of constraints than before in terms of space around
"// camel-k: " and "# camel-k: ". the Runtime allows to write
""//camel-k:" for instance. this can be improved in another task. it
wasn't the goal of this one.
- support space and tab character between options. The runtime is using
a regexp which seems to allow \r \n and \f but it will create invalid
files so we don't need that.


Signed-off-by: Aurélien Pupier <apupier@redhat.com>