FUSETOOLS2-721 - outline of Camel endpoints for XML
- it has kind of redundant imbrication for route/from but I preferred to
keep it as-is. It stays closer to Java outline. I think it is
interesting in outline, it allows to have different details. it is not
very nice with Ctrl+shift+O as it highlights the whole content route and
not just the from. That said, there is also the "from" highlighting only
the from line with the xml Document symbol which is usually also
installed. I do not clear advantages for one or the other option. So
picked the one easiest to maintain and better for the outline instead of
the Document symbol search which is a less visible feature.

Signed-off-by: Aurélien Pupier <apupier@redhat.com>