Fix logging mechanism
the file wasn't created. It seems that the java.io.tmpdir wasn't
replaced. So migrated to log4j2
There is a compatibility layer which allows to read log4 1 files but
well as I was at it, it also allows to use a more advanced mechanism -
rolling files and a better feedback on what is happening to investigate.

Signed-off-by: Aurélien Pupier <apupier@redhat.com>