FUSETOOLS2-1505 - cancel previously launched diagnostic when clearing
In case, a diagnostic was requested, then a clear request is coming
(through closing editor). If the diagnostic was taking some time, the
server was responding first the clear request and then the response for
the first diagnostic. So client will have the wrong last state
requested.
With this PR, it is ensuring that currently running request are
cancelled before clearing.

It was causing the test
CamelDiagnosticTest.testValidationErrorClearedOnClose() to be flaky

Signed-off-by: Aurélien Pupier <apupier@redhat.com>