FUSETOOLS2-1870 - avoid flakiness by checking timeout more precisely
The goal of the test is to ensure that the end-user receive the
completion feedback in less than a second. The timeout on the whole test
before was impacted by the time to setup the environment for the test.
Keeping a timeout at test level so that in case it hangs, it does not
hangs for several minutes.

Signed-off-by: Aurélien Pupier <apupier@redhat.com>