FUSETOOLS2-2182 - Completion for traits in VS Code tasks.json file
it was previously in VS COde Camel K codebase. With the move to Camel K
2, the list of traits is no more available through the CLI. So moving
this code in LSP which already has access to list of traits through the
CRD.
Note: the availability of completion is more permissive than previously
because I have not found an easy way to have access to JSon Node based
on position in the text in Java. There as a great utility in VS Code api
that is missing here. Given the low usage and impact it produces, i
think it is fine to live this way for now.

Signed-off-by: Aurélien Pupier <apupier@redhat.com>