FUSETOOLS2-2478 - completion for dependencies of Camel component in pom
there will be completion provided sometimes not at the exact right-place
but should be fine for now.

to have a good ones in this iteration the user will have to use the
correct runtime provider otherwise it will end up with not specific to
their runtime dependency. To improve in the future by helping to set the
runtime provider or to look to the content of the pom.xml

Signed-off-by: Aurélien Pupier <apupier@redhat.com>