FUSETOOLS2-2508 - Configure log to use a specific Camel LSP folder
otherwise when trying to walk files to delete old logs, depending on the
content of /tmp and the rights, an exception can be thrown. When this
exception is thrown, it is sent back to the Client in a way which is not
"correct" (I do not know why) and consequently, the client is sending a
shutdown request on the server!

It will also have the side effect to be more performant by not scanning
the full /tmp folder everyday or every time the log reach 10MB

Signed-off-by: Aurélien Pupier <apupier@redhat.com>