FUSETOOLS2-2529 - avoid NPE in FoldingRange
it sounds like the IDE is requesting a folding range (so places to place
little +/- to fold a part of the text) on a not opened document. I do
not understand what could be the purpose. Ignoring the request to avoid
NPE until we have more information for another use case (uri provided in
a different shape?). It will avoid the NPE and the missing
functionalities won't be critical.

Signed-off-by: Aurélien Pupier <apupier@redhat.com>